Conflicts:
	pom.xml
	src/main/c/build/common.sh
	src/main/c/build/linux-x86_64.sh
	src/main/c/configure.ac
	src/main/c/src/ConfigDescriptor.c
	src/main/c/src/DeviceDescriptor.c
	src/main/c/src/EndpointDescriptor.c
	src/main/c/src/EndpointDescriptor.h
	src/main/c/src/Interface.c
	src/main/c/src/InterfaceDescriptor.c
	src/main/c/src/LibUsb.c
	src/main/c/src/Makefile.am
	src/main/c/src/Version.c
	src/main/c/src/usb4java.h
	src/main/java/de/ailis/usb4java/libusb/ConfigDescriptor.java
	src/main/java/de/ailis/usb4java/libusb/Context.java
	src/main/java/de/ailis/usb4java/libusb/Device.java
	src/main/java/de/ailis/usb4java/libusb/DeviceDescriptor.java
	src/main/java/de/ailis/usb4java/libusb/DeviceHandle.java
	src/main/java/de/ailis/usb4java/libusb/DeviceList.java
	src/main/java/de/ailis/usb4java/libusb/EndpointDescriptor.java
	src/main/java/de/ailis/usb4java/libusb/Interface.java
	src/main/java/de/ailis/usb4java/libusb/InterfaceDescriptor.java
	src/main/java/de/ailis/usb4java/libusb/LibUsb.java
	src/main/java/de/ailis/usb4java/libusb/Transfer.java
	src/main/java/de/ailis/usb4java/libusb/Version.java
	src/main/resources/de/ailis/usb4java/libusb/linux-x86/libusb4java.so
	src/main/resources/de/ailis/usb4java/libusb/linux-x86_64/libusb4java.so
	src/test/java/de/ailis/usb4java/libusb/LibUSBDeviceTest.java
	src/test/java/de/ailis/usb4java/libusb/LibUSBTest.java

+++ src/main/java/de/ailis/usb4java/AbstractDevice.java
@@ -121,7 +121,7 @@ abstract class AbstractDevice implements UsbDevice
         for (int i = 0; i < numConfigurations; i += 1)
         {
             final ConfigDescriptor configDescriptor = new ConfigDescriptor();
-            final int result = LibUsb.getConfigDescriptor(device, i, 
+            final int result = LibUsb.getConfigDescriptor(device, (byte) i,
                 configDescriptor);
             if (result < 0)
             {
@@ -521,8 +521,8 @@ abstract class AbstractDevice implements UsbDevice
     {
         final DeviceHandle handle = open();
         final ByteBuffer buffer = ByteBuffer.allocateDirect(256);
-        final int result = LibUsb.getDescriptor(handle, LibUsb.DT_STRING, 0,
-            buffer);
+        final int result = LibUsb.getDescriptor(handle, LibUsb.DT_STRING,
+            (byte) 0, buffer);
         if (result < 0)
             throw new LibUsbException(
                 "Unable to get string descriptor languages", result);
